requirements.txt
click==8.1.3
Flask==2.2.2
Flask-WTF==1.1.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==0.21.0
Werkzeug==2.2.2
WTForms==3.0.1
